[BEAM-11383] Add runtime normalization for sdk.properties. (#13464)
This normalizes the sdk.properties file which has a build timestamp in it as well as the version and sdk version. For the purposes of the runtime classpath, these can be ignored so that tasks are not rerun if only the build timestamp is changed.
This should result in significantly higher cache hits as well as more up-to-date checks.

This also preserves timestamps for certain tasks for the same reason.

BEAM-11383